+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Changed
 
+- Simplify log messages for military zone verification
 - Docs: README.md update
 - Build: dependencies: upgrade to arma3-offline-map-lib 0.8.0